autoscale axes when map_extent is None
This commit changes the behavious of the None value for the map_extent formatoption. Now, map_extent=None triggers a call of the plotter.ax.autoscale method, whereas the old behaviour was to use the lonlatbox of the lonlatbox formatoption. This causes problems for regional files close to the map-boundary (see https://github.com/psyplot/psy-maps/issues/11).

The old behaviour can be retained by setting map_extent='data'